Add OpenGL function loading
Adds a function wrapper for loading the OpenGL functions, in practice this is not the most efficient way to do this as I have to explicitly call Load on each wrapper (The macro has eased some of this pain).
I suppose a hacky way around this would be to create a base class with a virtual "Load" function, which the wrappers inherit. Then have the wrappers register themselves to a temporary "Loader", which iterates the base pointers calling "Load".
I just wanted to have a go at loading the functions using some template programming.