2D Texture Slice Renderer
This is a rather large commit, but oh well. This adds my basic implementation of the 2D Texture Slice rendering algorithm from Real-Time Volume Graphics by Klaus Engel et al.
There are lots of optimizations that could be done to improve this - for a start using an index buffer - currently however, I think it is fairly easy to read and understand the algorithm code.